Update build/add-bundle script to add entries in correct place
When an entry is being added, regardless if is existing or new, the script will add it to the end of the entries list. Which causes if the version is lower than the last stable version to be in the wrong place and have wrong value of version in replaces field. The changes in this commit are fixing it

Signed-off-by: Ismail Ibrahim Quwarah <iquwarah@redhat.com>